added check for v4 vs v6 addresses
  * during testing of our cluster someone removed and the re-added
the v4 ip address from my k8s lb service.  this caused the IPs for
the service to be returned in a new order, IPv6 first, IPv4 second.
So I could no longer cheat and assume that the first addr would
always be the v4 address
  * I am still assuming that the service or pods only have a single
ipv4 or v6 address