added support to run tests from a single external client
  * multiple clients require changes to how result data is collated
as well as changes to the k8s test infra is set up.  For example,
need to add a second pod to the service because you can't send two
iperf clients against a single iperf server.  presumably the same
is true of statexec which synchronizes the start and end of the
client and server.